Lib\Cli - fix unlink pidFile

+++ Lib/Cli/cli.php
@@ -151,7 +151,9 @@ class Cli extends Element
   }
 
   public function removePid(){
-    unlink($this->pidFile);
+    if (is_file($this->pidFile)){
+      unlink($this->pidFile);
+    }
   }
 
   /**
